1.1.0
1.1.0

+++ csgo_showposition/ContainerHelp.py
@@ -2,6 +2,6 @@ from dataclasses import dataclass
 
 @dataclass
 class SpeedVector:
-    x_speed:int
-    y_speed:int
-    z_speed:int
+    x_speed:float
+    y_speed:float
+    z_speed:float
